EXXT5T8XX-5023: poweroff seems to fail in some uncertain situations
Looking at outputs from Mkarrman 2017-02-24, It seems that when the system
fails to suspend/poweroff correctly, when we use 'orderly_poweroff(1).
In the logs we have some output about a regulator  in LDM driver
(ca111) that wants to shut off one time to many, however,
we do pass this and get a message 'reboot: Power down' which is found in
reboot.c, function kernel_power_off.
This function will in turn disable interupts, stop processors etc, then
call pm_power_off

Not really sure what goes wrong, the symptom is that the camera is discharged
when being left in suspend.

But the output from Mkarrman log, shows that orderly_poweroff in faddev.c has
been called, and since 'reboot: Power down' is shown we have arrived in
function kernel_power_off before something fails, which is not shown.

To force the camera to power off, we call pm_power_off instead of orderly_poweroff.
The differences are that orderly_poweroff will try to use the systems shutdown command
to perform the shutdown, and if that fails, try to sync filesystems before calling pm_power_off
Running pm_power_off directly shuts down the pmic (da9063)

The filesystems have been synced when the system entered suspend
(line 'PM: Syncing filesysstems' in the log)

Not sure that we can assume that filesystems are stil synced when we resume,
particulary due to PM: resume of devices lines, and Restarting tasks, so It could be bad to
execute pm_power_off this late during the resume of the system!

Signed-off-by: Bo Svangård <bo.svangard@flir.se>