EXXT5T8XX-5023 Battery drain after night in standby
Earlier we have tried to force power off with pm_power_off, which did power off,
but problem with buzzing focus motor remained, due to the +4VA voltage being
always active on EVRO board, and when powering off the PMIC, this voltage ends
up latching the voltages and keeping the EVRO board active.
This was due to the FAD driver not having a shutdown function handling the case when
the system entered power_down mode

This patch adds a shutdown function to the fad driver, shutting down the system exactly like
the suspend function in the FAD driver.

Signed-off-by: Bo Svangård <bo.svangard@flir.se>