ec70x: Use fad_ninjago [LEIA-1167]
Use the general ninjago version for ec702 and ec701.
The fad driver uses a few regulators related to the lens
and focusing, and the system uses dummy regulators if
these are missing.
Introduced bHasFocusModule attribute and set ec702 to false
to avoid warnings about these regulators. ec101 and ec501
sets it to true in dts.